add magit-mode-setup, all but replacing magit-mode-init
Add new macro `magit-mode-setup' and use it where `magit-mode-init' was
previously used.  Keep the latter for compatibility and because we might
run into situations where the former does to much.

`magit-mode-setup' calls `magit-display-mode-buffer' to display and
select a buffer and then calls `magit-mode-init' to turn on the mode,
store refresh information and perform the first refresh.

It also makes sure that the arguments to `magit-mode-init' are evaluated
in the correct buffer - that is before switching to the mode buffer.
This helps prevent scope errors in the future.  (I am aware of at least
one instance when someone (me) has made exactly that mistake.)