move definitions of magit-save-index and magit-show
These functions do not belong in the "Diffs and Hunks" section.

`magit-show' has absolutely nothing to do with that, unfortunately
there no place at all where it would naturally belong.  Add it to
the new "Miscellaneous Commands" section.

Move `magit-save-index' to section "Ediff Support" as it is only
used there.

Also move the definitions of the related variables so that their
definitions remain located before their first use.

Move `magit-show-current-version' to "Ediff Support".

Move `magit-file-name' to "String and File Utilities" as it is also
used by `magit-buffer-file-name' added in the previous commit.