move definition of magit-init
Move definition of `magit-init' below that of `magit-status'.
The fact that the former may be used by the latter is the only
justification for it to be defined somewhere in the "Status Mode"
section at all.

The old location was just silly.  It was defined below the refresh
function and above the mode definition; and nowhere near that of
`magit-status'.