add new subsections to "Status Mode" section
The "Status Mode" toplevel section contains various definitions that do
not strictly belong there.  Because there also are no other places that
would make complete sense and defining separate toplevel sections for
these definitions would also not be appropriate they are kept in the
current section.

The hope is that eventually a natural place for these definitions will
evolve.  Until then at least split the misplaced stuff into dedicated
subsections.  This can be done without moving any code around (except
for `magit-init' which was moved in the last commit).

Add a new subsection "Save File Buffers".  Because it right in the
middle of the existing "Status Refresh and Commands" subsection that is
renamed to "Status Refresh" and a new subsection "Status Commands" is
added.

Creating new toplevel sections would also be an option but that would
require that code be moved around (because "Status Commands" really does
belong in the "Status Mode" section).

In summary: "Save Buffers" and "Read Repository" should not really be
located below "Status Mode" but having them their is less bad than the
alternatives.