magit-run*: don't use pty in all cases where we shouldn't
As someone noted a while back on the mailing list we never use a
pty when writing to stdin of a "synchronous" git process but we
might use a pty when git is running asynchronously.  (Actually
even in the "synchronous" case git is running asynchronously, we
just poll until it is done be fore returning.)

This might fix #255.