replace global magit-process-client-buffer with local variable
Instead of setting `magit-process-client-buffer' globally pass the
buffer that was current when `magit-run*' was called to the functions
that need to know about its value.  These are the process sentinels.

Because process sentinels take exactly two arguments we create a partial
application of `magit-process-sentinel' (and `magit--bisect-run-sentinel')
and use the returned function as actual sentinel.

Also limit the use of the global variable `magit-process-buffer-name',
rename some of `magit-run*'s internal variables and change were these
variables are set.