remove unused magit-run-git-async-with-input
No internal or external libraries use this function.  The old commit
workflow used it but even if we restore parts of that we should not
restore this function (we could just use `magit-run-git*' directly).
I would like to avoid advertizing this usage of `magit-run*' (which
this function is just a wrapper around).