define magit-git-{string,lines} more directly
Instead of returning the result of passing the value of
`magit-git-output' to `magit-trim-line'/`magit-split-lines' use
`call-process' and extract the first line / the lines directly.

Add doc-strings because we need to explain how empty lines are handled.

Change how `magit-git-lines' deals with empty lines.  Previously the
return value could contain empty lines but if the first line was such a
line then that empty string was removed, but no other.  Now all empty
lines are ignored.