move "Submoduling" subsection into first "Acting" section
For historic reasons there are currently two top-level "Acting"
sections.  While it would be best to merge them, there is no need to do
that right away - especially as future changes will likely influence how
that should be done.

However "Submoduling" should be moved into the first instead of the
second "Acting" section now because it is the only remaining subsection
that contains commands that are used in a key-mode popup that is still
located in the second "Acting" section.