magit-ignore-item{,-locally}: allow editing ignore string
When used with a prefix argument `magit-ignore-item{,-locally}' now
let the user edit the ignore string.  `magit-ignore-file' already had an
optional argument to trigger the editing but it was not used.

Also merge `magit--ignore-item' into `magit-ignore-item', rename
`magit-ignore-file's argument EDIT-IGNORE-STRING to just EDIT, and when
prompting for the file or pattern use a prompt that actually sais either
of the two is valid.