add magit-get-tracked-branch
Add `magit-get-tracked-branch' replacing `magit-remote-branch-for'.
The new function is a bit to complex (contributions welcome) but it
is better than the old which just assumed that there can only be one
fetch rule: "+refs/heads/*:refs/remotes/<name>/*".

`magit-get-tracke-branch' doesn't have that limitation.  Additionally
it does only return a branch name instead of a full branch path if a
a non-standard path is used.  The new optional PRETTY argument can be
used to return the formatted name `magit-format-ref' would return.

Refactor `magit-insert-status-remote-line' to make use of the new
function.