refactor magit-insert-{unpulled,unpushed}-commits
This results in few additional calls to `git config' but this does not
reduce performance in a noticable way.  On the other hand it prevents
converting some data from A to B and then back to A.  Remove the
function `magit-remote-branch-name' which was used to make this leaky
abstraction less visible.