add magit-diff-abbrev-arg and magit-log-cutoff-length-arg
Add `magit-diff-abbrev-arg' and `magit-log-cutoff-length-arg' so we
don't have to repeatably construct the respective Git arguments based on
the values of `magit-sha1-abbrev-length' and `magit-log-cutoff-length'.

Also always use "-1" not "--max-count=1" or ("-n" "1") to turn `git log'
into a command that shows information about exactly one commit.  Git
should really have such a command, which is as powerfull as `git log -1'.
Well there is not command `git log-1' but `git log -1' is close enougth;
consider it a command, `git log --max-count=1' is harder to see as ac
command.