fix magit-refresh-reflog-buffer inconsistencies
Rename the argument REV (revision) of `magit-reflog' and
`magit-refresh-log-buffer' to REF (reference).  The reflog
deals with branches not with arbitrary revisions.

Remove `magit-refresh-reflog-buffer's ARGS argument.  Its name
was wrong it wasn't a list but an atom.  The proper name would
have been something along the line RESOLVED-REF.  Remove the
argument entirely because we might as well evaluate
`(magit-rev-to-git ref)' inside `magit-refresh-reflog-buffer'
instead of `magit-reflog'.