magit-item-highlight: use definition similar to that of highlight
When `magit-diff-use-overlays' is non-nil set the default value of face
`magit-item-highlight' to something similar to the default value of
`highlight'.  Do not inherit from `highlight' but instead copy the
value (with some redundancy removed), because we want to use a blue
instead of green background in the "very few colors" case.  Green is
already used in diffs for "added lines".  Also this potentially avoids
problems with themes.