Move "Commit Mode" heading, rename "Logs{ and Commits => }"
Move the beginning of the "Commit Mode" top level section up to include
definitions that were previously part of the "Logs and Commits" section.
Rename that to just "Logs" as it no longer contains code related to the
commit buffer.  (Having a section that contains both "log" and "commit"
related code didn't make sense in the first place.)

The order of the code inside the extended "Commit Mode" section is
unfortunate but we can live with it for a while.