magit.el: add preliminary subsections
Add preliminary subsections to the top-level sections "Revisions and
Ranges", "Diffs and Hunks", "Logs", and "Branch Manager Mode".  Wrap
the headings with parenthesis to indicate their status and also add
comments to that effect.

I have added subsections in the past only after/when rearranging the
containing top-level section.  But because I don't want to do this for
all these sections now I am adding these preliminary subsections now
without at the same time bringing the subsections and the contained
definitions into a reasonable order.

In my cleanup efforts I am currently focusing on removing *other* code
smells.  E.g. by identifying utilities that need to generalized to be
used outside of the scope they were originally written for (and in the
process removing narrow minded simplifications) and removing functions
only used by one caller (because it wouldn't make sense to improve and
then use them elsewhere).

Given my tiny brain it helps when code is stored in little labeled
boxes.  In the past I *have* used such temporary headings like those
being added in this commit.  I just never committed them and that let
to a lot of stashing and other laborious trickery.  Not this time: this
commit gives me a structure to work in and which is to be improved.

So yes, it's ugly and doesn't make for a good first impressions but for
all that it is worth doing so helps me fix the underlying causes of the
disorganization more quickly.