add new top-level "Font-Lock" section
"Move" everything related to font-locking into new top-level "Font-Lock"
section.  This includes all the header-related font-locking previously
located in the "Headers" section and the only function that was actually
moved: `git-commit-font-lock-diff'.  According to Git what was moved is
the unrelated `git-commit-mode-syntax-table' but that is due to patch
minimizing gone wrong.