remove magit-flow.el for now
Remove the `magit-flow.el' for the time being.  Another
implementation that implements the functionality described at
http://jeffkreeftmeijer.com/2010/why-arent-you-using-git-flow
will be added at a later time.

This version is removed because it depends on external scripts
that implement this workflow (https://github.com/nvie/gitflow),
which happens to be unmaintained.  It will actually be easier
to rewrite this in directly elisp than wrap around a bugged
external implementation (and eventually having to fork and
maintain that).  Also in most distributions these scripts are
not available through the native package manager.

More importantly though this version has to be removed because
it is very incomplete.  Having this version in the Magit repo
will raise expectations we cannot fulfill at this time; for a
elisp-only implementation to be feasible some existing mess in
the Magit code base has to be cleaned up first, and extenting
the wrapper around the shell scripts would be a waste of time
considering the goal of not relying on these eventually.

See issue #881 for discussion.