`magit-bisect-run': properly set `buffer' and `process'
In an effort to limit the usage of `magit-process-buffer-name',
commit 54f2e7c2 modified `magit-bisect-run' to set `process' and
`buffer' based on the return value of `magit-run-git-async'.

However, `magit-run-git-async' uses `magit-run-git*', which uses
`magit-run*', which returns its `successp' variable, which is either
t or nil, *not* a process object, so `magit-bisect-run' fails when
it tries to set `buffer' by running `process-buffer' on `process':

  Debugger entered--Lisp error: (wrong-type-argument processp t)
  [...]

Fixes #891.

Signed-off-by: Pieter Praet <pieter@praet.org>