`magit-mark-item': fix toggling mark in status buffer
Running `magit-mark-item' on a marked commit is supposed to unmark it.

Since the value stored in `magit-marked-commit' is a string, doing
comparison using `eq' can't possibly work, so use `equal' instead.

BTW, for some odd reason (which I don't feel like investigating right
now), the previous state of affairs *did* in fact *work* in the log
buffer (as opposed to the status buffer).  Doesn't matter anymore
though.  Now it works in both of them...

Fixes #898.

Signed-off-by: Pieter Praet <pieter@praet.org>