magit-svn-get-ref-info: insert complete commit information
Insert the complete log output about the found commit into the
buffer.  Previously only the first line was inserted but the
line we later search for is actually the last line in of the
commit message.

The way this information is retrieved could be improved but I
have never used svn so this will have to do.  I also could not
test this, but it should work.

Close #884.

Reported-by: Yasushi Abe <yasushi.abe@gmail.com>