`magit-git-string': optimize
Using a stock configuration, running `magit-status' results in 27 calls
to `magit-git-string', so the less it does, the better.

Checking whether the buffer contains anything can be done in a single call
(`buffer-size') instead of two (`point-min', `point-max'), and `point'
should only be moved once we're sure the buffer contains anything.

Signed-off-by: Pieter Praet <pieter@praet.org>