add missing magit-{*}-buffer-name variables, change doc-strings
Add the missing `magit-{*}-buffer-name' variables so that such a
variable exists for each Magit mode.  Two such variables have to
be declared to suppress compile warnings, so be it.

Also add doc-strings to existing variables that lack it and give
all such doc-strings the same form.

Also rename `magit{--cherry => -cherry}-buffer-name'.