teach magit-parse-log-line about unpushed/unpulled logs
Extend `magit-parse-log-line' to support the log style used to display
unpushed and unpulled logs.  The words used to describe the list of
unpushed and unpulled logs hide the fact that these are essentially
ranged logs

Now that we have noticed the `magit-wash-unpushed-or-unpushed' function
which essentially duplicates `magit-wash-log' can be removed, hurray.