move `magit-ref-get-label-color' into new "Reference Labels" section
Move the `magit-ref-get-label-color' function and all support functions
and variables into the new "Reference Labels" subsection of "Revision
and Ranges" (which is also, or mainly about References, but that's a
different story).

The support symbol names of support functions and variables sport the
"log" substring and indeed the functionality provided by the moved code
is *currently* only used for logs.  However "colorizing a reference
according to its type" is also needed elsewhere.  Currently these other
places use different less flexible implementations, but that should be
changed.  So move as a first step toward using this (or rather a
generalized variant thereof) move it to a "common" place.

Also move the `magit-log-format-reflog' into the "Reflog Mode" section;
*not* to the same place as `magit-ref-get-label-color'!  These functions
do something similar and are both used by `magit-format-log-line' but
they actually are quiet different.

`magit-log-format-reflog' is about colorizing the "actions" in reflogs,
not about doing that to references, so it does not belong near code that
deals with references.  It could be left were it was previously defined,
since that is also where the only call to it is located (inside
`magit-format-log-line').  However because it is strictly only related
to reflogs the respective section is the perfect place.