make variable magit-log-format constant
Make variable `magit-log-format' constant.  Unlike the
`magit-reflog-format' variable removed in the previous commit
it is actually used multiple times, so even though it likely
will never change we keep the name stuck onto this git argument.

Because it is a boring detail, move it to the bottom of its
section were only truly curious readers will see it.