move definition of magit-log-cutoff-length-arg
Move the definition of the function `magit-log-cutoff-length-arg'.
Since it has variable-like characteristics keep it in the "variable"
section but move it to the end its.  This way only the curious reader
will see it - starting out a "variable" section with a function is a
bit ... confusing.

Also this (and variable `magit-log-format') do not /really/ belong in
this section at all.  But I cannot find a better place to stick them
at them moment.  Another reason to move them out of sight.