move definition of magit-create-log-buffer-sections
Move the definition of the macro `magit-create-log-buffer-sections' from
the "Log" section to the "Section Creation" section.  Both places make
sense, but I prefer the new one because now all the functions and macros
used to create sections are in one place.  Also moving this macro means
that the "Log" section now only contains things /directly/ related to
washing.

Also move the definition variable `magit-log-count', as it is set by the
macro.

Lose the doc-strings, they are not good.