rename magit-{svn-insert => insert-svn}-{status- => }remote-line
For consistency with other such inserters in magit.el switch "svn" and
"insert".  Also remove the "status-" part.  While this is intended for
use in status buffer, that is not mean we need to keep that part.
Actually it should also be removed for some other inserters that contain
it, since they can be used outside the status buffer.