magit.el: fix Emacs compat function versions
Specify in which commits the functions were added to Emacs, and correct
the version numbers for `string-prefix-p' and `string-match-p'.

Note that we try to ensure compatibility with Emacs 23.2 and up, so the
compatibility functions for `string-prefix-p' (added in 23.2 instead of
23.3) and `string-match-p' (added in Emacs 23.1 instead of 23.3) are
nothing but clutter, meaning we should probably sacrifice them to Lord
Lambda so we can recycle the parens.

Signed-off-by: Pieter Praet <pieter@praet.org>