magit-file-log: always allow selecting the file
Allow users to select the file for which to show the log.  Previously
that was only possible when using the popup, but when `magit-file-log'
was invoked directly from a file-visiting buffer, that file was used
unconditionally.  Showing the log of that file certainly is the most
common case - now it is one key press further away.