magit-blame-{prev,next}-chunk
Always move to the previous/next chunk from *point*.  Previously that
was done interactively but non-interactively the POS argument had to
be provided.  Which is wrong for two reasons:  we did not use it
non-interactively, so even if this was a good idea in the first place
it would be of no use.  And then there is no reason for a movement
command to be have "move to next thing from point, except when I tell
you to move from somewhere else".  If one wants that - especially in
non-interactive use, one can just go to that other place *before*
going to the next thing,