Add option to omit merge commits from status buffer
Some users think that merge commits in the status buffer are somewhat
superfluous since the actual commits being merged together should also
show up in the list. Add an option to omit them.

Since this is a matter of personal preference leave the default behavior
unchanged, which is to show merge commits in the status buffer.