remove unnecessary function magit-hunk-item-diff
The parent section of a `hunk' is always a `diff'.  Function
`magit-hunk-item-diff' is therefor unnecessary.  All it did was assert
that this is indeed the case.  By removing this assert we only risk that
a user reports a harder to diagnose error message, but if that ever
happens it should be doable.  Besides I cannot remember that there ever
was a report Magit complained "Huh?  Parent of hunk not a diff".