magit-wash-diffstat: remove comment and commented code
The comment was added in commit 293a7804b1b0b15614c0e4c826f9e70059e872ff
the commented code consists of the "removed" code and, so it seems, some
leftover experimental code.

The comment would have made a fairly good commit message, but as a
comment it does not fly with me.  Leaving the old code in the file but
comment it...