magit-wash-diff-section: replace a use of let* with let
In `magit-wash-diff-section' replace one unnecessary use of `let*' with
`let'.  This avoids confusion in the future about why `let*' was used in
the first place.  To keep the diff small replace "*" with a space,
instead of removing it.  This avoids having to reindent the whole thing,
resulting in a rather larger diff, for no good reason.