make variable magit-ignore-unmerged-raw-diffs unnecessary
Instead of let binding `magit-ignore-unmerged-raw-diffs' inside
`magit-staged-changes' call `magit-wash-raw-diffs' with its new STAGED
argument, which is then passed on to `magit-wash-raw-diff' which happens
to be the only place where this information is relevant.  (And
`magit-staged-changes' is the only case where we have to toggle this
switch; hence the argument name "STAGED".