remove magit-indentation-level
The default value was 1, the dynamically bound value always was 2; so we
might as well make it static.  However the correct value would actually
be 1 not 2.

Why didn't we notice so far? `magit-wash-typechange-section' is dead
code by now, it is never actually called.  (This has to be fixed and
is a likely cause of issue #224.)