collect all status section inserters in one place
All of the sections created by these inserters are used in the "status"
buffer.  They can also be used in other Magit buffers but that doesn't
change the fact they are "status" sections.

Some of these inserters were defined as part of the "Status Mode"
toplevel section, while others where defined next to the lowlevel
functions they make use of.  There are two ways to make that
consistent... this commit implements on of them.

Because these inserters can also be used by other Magit modes besides
`magit-status-mode' create a new toplevel section "Status Sections"
instead of placing inside a subsection of "Status Mode".  Actually this
is done by promoting the existing subsection "Status Mode">"(sections)"
to a topsection; so the definitions that were already located there do
not have to be moved.

Because after moving `magit-insert-untracked-files' the toplevel section
"Untracked Files" only contains one function `magit-wash-untracked-files'
move that along with the former (which happens to be its only caller).

The definitions of the inserters are in the same order as the sections
are inserted in the "status" buffer.  Unfortunately that means that the
more noisy ones come first.  But that's just because nobody has bothered
cleaning them up yet.

Actually that is not quiet true.  The "line" sections are defined after
the "real" (multi-line and collapsible) sections.  But that is no
problem because the section is split into respective subsections "Line
Sections" and "Real Sections".