move magit-section pseudo getters into the struct section
When a `magit-section' struct represents a `diff' or `diffstat' section
more information has to be stored inside the struct than for other
section types.  Currently this is done by storing a list in the `info'
slot.

They right way to implement this would be to add additional slots to
the `magit-section' struct and just not use them for the other types,
that don't need them.

But because the `info' slot is used to uniquely identify specific
section this is not currently possible.  Doing so would require complex
refactoring elsewhere. I have already attempted to do so, it requires
changes all over the place.

Instead, for now, move the `diff` and `diffstat` specific pseudo slot
getters below the definition of the struct.