don't require saveplace
If the user hasn't setup saveplace in their config, requiring saveplace
causes them to get a useless .emacs-places in the home directory. To
avoid triggering a compilation warning, we check if `save-place' is
bound.

f5b36db4 had changed from calling `toggle-save-place' to setting
`save-place' but still checked if `toggle-save-place' was fbound,
830b538b then required saveplace to quiet the compiler.